requirements.txt
ultralytics>=8.0.196
albumentations>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.1
numpy>=1.24.3
pandas>=2.0.3
pyyaml>=6.0
Pillow>=10.0.0
seaborn>=0.12.2
tqdm>=4.65.0
